requirements.txt
pykrx>=1.0.47
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
anthropic>=0.25.0
apscheduler>=3.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
slack-sdk>=3.27.0
streamlit>=1.35.0
matplotlib>=3.8.0
setuptools>=75.0.0
